Fix linkage of utilities in the Makefile.
First of all, rpath was being set to NULL, which makes it unsafe (empty
rpath is the same as ".", letting the loader look for libraries in the
current work directory); the obvious mistake here was to use $(LIB) rather
than $(LIBDIR) (did it change over time?).

But make RPATH optional, by moving it to its own macro definition, this
allows for packagers to not use rpath at all, as it's usually not necessary
when installing in the default library paths (such as /lib).

Also, move the -L. flag at the top of the link command, so that
user-provide library search paths won't cause another libkeyutils to be
linked against.

Cc: Robin Johnson <robbat2@gentoo.org>
Signed-off-by: Diego Elio Pettenò <flameeyes@gmail.com>
Signed-off-by: David Howells <dhowells@redhat.com>