Build: Extract version info when building rather than defining it redundantly
Make the Makefile extract version information from the specfile and the library
version script when building rather than defining it redundantly.

The specfile provides the current package version and the library version
script provides the current API version.

Also add a 'make rpm' option to build a RPMs from the HEAD version of the GIT
tree that the keyutils package is in.

Signed-off-by: David Howells <dhowells@redhat.com>