keyctl: Allow multiple lines in the commands[] table with same name
Allow multiple lines in the commands[] table with same name by setting the
action function pointer to NULL for the second and subsequent entries.  This
prevents the partial matcher from proclaiming ambiguity for short versions of
commands that have multiple lines ('session' and 'purge').

Also alphabetically order the commands table.

Signed-off-by: David Howells <dhowells@redhat.com>