TEST: Unlink can now take a single argument
Remove the unlink argument validity check that checks that just one argument
fails as it's now valid to pass a single argument.

Add a test for the unlink with single argument case (which searches the process
keyrings for all instances of the specified key and attempts to unlink them
all).

Signed-off-by: David Howells <dhowells@redhat.com>