TEST: fix permissions on created keys and keyrings
Recent upstream kernel commit:

    commit 96b5c8fea6c0861621051290d705ec2e971963f1
    Author: David Howells <dhowells@redhat.com>
    Date:   Tue Oct 2 19:24:56 2012 +0100
    KEYS: Reduce initial permissions on keys

reduced initial permissions and some tests are now hitting EACCES
because they don't get to use the 'possessed' key permissions, but
must instead use the user/group/other permissions only.

This particularly affects the recursion test because the test for possession
hits the recursion limit and we don't see keys of that depth as being
possessed.  To fix this, the keyrings in the recursion test are given full
user access and the eighth keyring is created elsewhere and linked in.

Signed-off-by: Jan Stancek <jstancek@redhat.com>
Signed-off-by: David Howells <dhowells@redhat.com>