keyctl: Fix command table searching
In the keyctl program, when searching the command table, don't trust that an
overlong argument name won't cause memcpy() to segfault.  This is unlikely to
be a problem as any efficient memcpy() is going to start scanning from the
lowest address, given that it has to report on the lowest-addressed difference
if there is one.

Reported-by: Al Viro <viro@zeniv.linux.org.uk>
Signed-off-by: David Howells <dhowells@redhat.com>