TEST: Insert a delay in the key invalidation test to allow for gc delays
The key invalidation test may fail because the check for a key having been
invalidated may occur too quickly after the call to invalidate the key.

The problem is that whilst the key is marked invalidated immediately, the
subsequent process of garbage collecting the invalidated key isn't synchronous
and may be delayed as the keyrings gc is merely scheduled to be run at some
later time.

Stick a small delay in there to give the gc a chance to run.

Bugzilla-entry: https://bugzilla.redhat.com/show_bug.cgi?id=1072798
Reported-by: Jan Stancek <jstancek@redhat.com>
Signed-off-by: David Howells <dhowells@redhat.com>