TEST: Allow skipping tests requiring installation
This is useful for running tests exlusively against the build result.

Modifing PATH and LD_LIBRARY_PATH is sufficient for most test cases, but in the
case of tests for the request2 and prequest2 subcommands, the kernel looks
for /sbin/request-key. Testing the build result would therefore require
installing it, but this is not always desirable, for example on build daemons.

Signed-off-by: Christian Kastner <debian@kvr.at>
Signed-off-by: David Howells <dhowells@redhat.com>